Se cambió prácticamente todo ya que básciamente muchas cosas se pasaban por alto. Por ejemplo aún se estaban utilizando estructuras nativas de java, entonces se implementaron (para litralmente todo) estructuras genéricas propias, para así literalmente generalizar todo. Aparte de eso, para las conexiones y aristas se manejaban métodos que retornaban datos tipo int, esto es un problema ya que los algoritmos pierden presición por legalmente los decimales. Ya me dió pereza seguir

+++ src/estructuras/NodoHeap.java
@@ -3,7 +3,7 @@ package estructuras;
 public class NodoHeap {
 	
 	int vertice;
-	int prioridad;
+	double prioridad;
 	
 	public NodoHeap(int vertice, int prioridad) {
 		this.vertice = vertice;
